Move operation objects to lib/operators/index.js
